// frontend-old/src/types/character/basic.ts
export interface Basic {
  date: string;
  character_name: string;
  world_name: string;
  character_gender: string;
  character_class: string;
  character_class_level: string;
  character_level: number;
  character_exp: number;
  character_exp_rate: number;
  character_guild_name: string;
  character_image: string;
}